fix: to modularize beforeEnter for User
