+++ src/modules/router-auth.js
@@ -1,0 +1,50 @@
+import axios from "axios";
+import crypto from "crypto-js";
+
+export default {
+  authHeader: {
+    "access-token": "",
+    client: "",
+    uid: "",
+  },
+  guardAccessToUser(Vue, to, from, next) {
+    if (Vue.$cookies.isKey("auth-header")) {
+      const decryptedAccessToken = crypto.AES.decrypt(
+        Vue.$cookies.get("auth-header")["access-token"],
+        Vue.prototype.$encryptKey
+      ).toString(crypto.enc.Utf8);
+      const decryptedClient = crypto.AES.decrypt(
+        Vue.$cookies.get("auth-header")["client"],
+        Vue.prototype.$encryptKey
+      ).toString(crypto.enc.Utf8);
+      const decryptedUid = crypto.AES.decrypt(
+        Vue.$cookies.get("auth-header")["uid"],
+        Vue.prototype.$encryptKey
+      ).toString(crypto.enc.Utf8);
+      this.authHeader["access-token"] = decryptedAccessToken;
+      this.authHeader["client"] = decryptedClient;
+      this.authHeader["uid"] = decryptedUid;
+      axios
+        .get("/api/v1/auth/validate_token", {
+          headers: this.authHeader,
+        })
+        .then((res) => {
+          console.log(res);
+          if (res.data.data.id == to.params.id) {
+            next();
+          } else {
+            throw "認証失敗";
+          }
+        })
+        .catch((err) => {
+          console.log(err);
+          next(from);
+        });
+    } else {
+      next(from);
+    }
+  },
+  // guardSourAuth() {
+  // write later (to 404 error page?)
+  // }
+};
