+++ src/router/index.js
@@ -1,18 +1,11 @@
 import Vue from "vue";
 import VueRouter from "vue-router";
-import LemonSoursIndex from "@/components/pages/LemonSoursIndex.vue";
-import LemonSour from "@/components/pages/LemonSour.vue";
-import User from "@/components/pages/User.vue";
-import axios from "axios";
-import crypto from "crypto-js";
+import LemonSoursIndex from "@/components/pages/LemonSoursIndex";
+import LemonSour from "@/components/pages/LemonSour";
+import User from "@/components/pages/User";
+import RouterAuth from "@/modules/router-auth";
 
 Vue.use(VueRouter);
-
-const authHeader =  {
-  "access-token": "",
-  client: "",
-  uid: "",
-}
 
 const routes = [
   {
@@ -30,42 +23,8 @@ const routes = [
     name: "User",
     component: User,
     beforeEnter: (to, from, next) => {
-      if (Vue.$cookies.isKey("auth-header")) {
-        const decryptedAccessToken = crypto.AES.decrypt(
-          Vue.$cookies.get("auth-header")["access-token"],
-          Vue.prototype.$encryptKey
-        ).toString(crypto.enc.Utf8);
-        const decryptedClient = crypto.AES.decrypt(
-          Vue.$cookies.get("auth-header")["client"],
-          Vue.prototype.$encryptKey
-        ).toString(crypto.enc.Utf8);
-        const decryptedUid = crypto.AES.decrypt(
-          Vue.$cookies.get("auth-header")["uid"],
-          Vue.prototype.$encryptKey
-        ).toString(crypto.enc.Utf8);
-        authHeader["access-token"] = decryptedAccessToken;
-        authHeader["client"] = decryptedClient;
-        authHeader["uid"] = decryptedUid;
-        axios
-          .get("/api/v1/auth/validate_token", {
-            headers: authHeader,
-          })
-          .then((res) => {
-            console.log(res);
-            if (res.data.data.id == to.params.id) {
-              next();
-            } else {
-              throw "認証失敗"
-            }
-          })
-          .catch((err) => {
-            console.log(err);
-            next({ path: "/lemon_sours" })
-          });
-      } else {
-        next(from)
-      }
-    }
+      RouterAuth.guardAccessToUser(Vue, to, from, next);
+    },
   },
 ];
